app_design/requirements.txt
aio-pika==9.3.1
aiormq==6.7.7
alembic==1.12.1
annotated-types==0.6.0
anyio==3.7.1
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.104.1
greenlet==3.0.1
h11==0.14.0
idna==3.6
Mako==1.3.0
MarkupSafe==2.1.3
httptools==0.6.1
multidict==6.0.4
pamqp==3.2.1
packaging==23.2
pika==1.3.2
psycopg2==2.9
pydantic==2.5.2
pydantic-settings==2.1.0
pydantic_core==2.14.5
python-dotenv==1.0.0
pytest==7.4.3
opentelemetry-distro==0.33b0
opentelemetry-instrumentation-fastapi==0.33b0
opentelemetry-exporter-otlp-proto-grpc==1.12.0
sniffio==1.3.0
SQLAlchemy==2.0.23
sqlalchemy_utils==0.41.1
starlette==0.27.0
typing_extensions==4.8.0
urllib3==2.1.0
uvicorn==0.24.0.post1
websockets==12.0
yarl==1.9.3
